+++ src/core/distributions/statistics/Dist_cat.cpp
@@ -89,9 +89,9 @@ const MemberRules& Dist_cat::getMemberRules( void ) const {
 
 
 /** Get the number of states in the distribution */
-int Dist_cat::getNumStates( void ) const {
+size_t Dist_cat::getNumStates( void ) const {
 
-    return int( static_cast<const Simplex*>( getValue("m") )->getElementsSize() );
+    return static_cast<const Simplex*>( getValue("m") )->getElementsSize();
 }
